# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentSnapshot": {
        "additionalProperties": false,
        "description": "创建本场角色时保存的模板快照。\n\n模板后续改动**不影响**历史与已有会话；同时记录模板来源 ID。",
        "properties": {
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "initial_goal": {
            "title": "Initial Goal",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "persona": {
            "title": "Persona",
            "type": "string"
          },
          "private_background": {
            "title": "Private Background",
            "type": "string"
          },
          "source_template_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Source Template Id",
            "type": "string"
          },
          "speech_style": {
            "title": "Speech Style",
            "type": "string"
          }
        },
        "required": [
          "name",
          "persona",
          "speech_style",
          "initial_goal",
          "private_background",
          "source_template_id",
          "captured_at"
        ],
        "title": "AgentSnapshot",
        "type": "object"
      },
      "AgentSpecRequest": {
        "additionalProperties": false,
        "description": "创建场景时对一名本场角色的要求。",
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "template_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "template_id"
        ],
        "title": "AgentSpecRequest",
        "type": "object"
      },
...
      "AnalysisCapability": {
        "additionalProperties": false,
        "description": "分析能力开关状态，用于健康检查与界面提示。",
        "properties": {
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "external_package_installed": {
            "default": false,
            "title": "External Package Installed",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "AnalysisCapability",
        "type": "object"
      },
      "AnalysisCapabilityView": {
        "additionalProperties": false,
        "description": "分析能力状态（供界面决定是否可用与如何解释）。",
        "properties": {
          "capability": {
            "$ref": "#/components/schemas/AnalysisCapability"
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          }
        },
        "required": [
          "scene_id",
          "capability"
        ],
        "title": "AnalysisCapabilityView",
        "type": "object"
      },
      "AnalysisCreateRequest": {
        "additionalProperties": false,
        "description": "请求分析：选定一名本场角色与若干**已提交公开材料**的序号。",
        "properties": {
          "agent_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "material_seqs": {
            "items": {
              "type": "integer"
            },
            "maxItems": 200,
            "title": "Material Seqs",
            "type": "array"
          }
        },
        "required": [
          "agent_id"
        ],
        "title": "AnalysisCreateRequest",
        "type": "object"
      },
      "AnalysisListView": {
        "additionalProperties": false,
        "description": "分析记录列表 + **分开呈现**的操作数与实际模型请求数（PRD 5.3）。",
        "properties": {
          "analysis_requests_used": {
            "minimum": 0.0,
            "title": "Analysis Requests Used",
            "type": "integer"
          },
          "capability": {
            "$ref": "#/components/schemas/AnalysisCapability"
          },
          "max_analysis_requests": {
            "minimum": 0.0,
            "title": "Max Analysis Requests",
            "type": "integer"
          },
          "operations_total": {
            "minimum": 0.0,
            "title": "Operations Total",
            "type": "integer"
          },
          "provider_attempts_total": {
            "minimum": 0.0,
            "title": "Provider Attempts Total",
            "type": "integer"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/AnalysisRecordView"
            },
            "title": "Records",
            "type": "array"
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          }
        },
        "required": [
          "scene_id",
          "capability",
          "operations_total",
          "provider_attempts_total",
          "analysis_requests_used",
          "max_analysis_requests",
          "records"
        ],
        "title": "AnalysisListView",
        "type": "object"
      },
...
      "AnalysisRecordView": {
        "additionalProperties": false,
        "description": "一条分析记录（含报告与计数，界面据此区分五状态）。",
        "properties": {
          "agent_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "analysis_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Analysis Id",
            "type": "string"
          },
          "behavior_description": {
            "default": "",
            "title": "Behavior Description",
            "type": "string"
          },
          "context": {
            "default": "",
            "title": "Context",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "degradation_flags": {
            "items": {
              "type": "string"
            },
            "title": "Degradation Flags",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "material_seqs": {
            "items": {
              "type": "integer"
            },
            "title": "Material Seqs",
            "type": "array"
          },
          "materials": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Materials",
            "type": "array"
          },
          "provider_attempts": {
            "minimum": 0.0,
            "title": "Provider Attempts",
            "type": "integer"
          },
          "report": {
            "$ref": "#/components/schemas/AnalysisReport"
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "analysis_id",
          "scene_id",
          "agent_id",
          "status",
          "provider_attempts",
          "created_at",
          "report"
        ],
        "title": "AnalysisRecordView",
        "type": "object"
      },
      "AnalysisReport": {
        "additionalProperties": false,
        "description": "分析结果。\n\n必须区分 normal／blocked／degraded／failed／disabled 五种状态；保留\n``degradation_flags``，不把返回 JSON 当成分析成功（PRD 6.2）。结果不进入\n角色记忆、不改变人设、不触发调度。",
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "alternative_explanations": {
            "items": {
              "type": "string"
            },
            "title": "Alternative Explanations",
            "type": "array"
          },
          "behavior_labels": {
            "items": {
              "type": "string"
            },
            "title": "Behavior Labels",
            "type": "array"
          },
          "degradation_flags": {
            "items": {
              "type": "string"
            },
            "title": "Degradation Flags",
            "type": "array"
          },
          "disclaimer": {
            "default": "",
            "title": "Disclaimer",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "mechanisms": {
            "items": {
              "type": "string"
            },
            "title": "Mechanisms",
            "type": "array"
          },
          "provider_attempts": {
            "default": 0,
            "minimum": 0.0,
            "title": "Provider Attempts",
            "type": "integer"
          },
          "scene_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "status"
        ],
        "title": "AnalysisReport",
        "type": "object"
      },
...
      "AnalysisStatus": {
        "description": "五种结果状态必须区分（PRD 6.2）。",
        "enum": [
          "NORMAL",
          "BLOCKED",
          "DEGRADED",
          "FAILED",
          "DISABLED"
        ],
        "title": "AnalysisStatus",
        "type": "string"
      },
...
      "Budget": {
        "additionalProperties": false,
        "description": "每场预算上限（PRD 5.3）。\n\n两项独立计数；创建前可在服务端允许范围内修改，开始时锁定（``locked_at``）。",
        "properties": {
          "locked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked At"
          },
          "max_analysis_requests": {
            "default": 4,
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Max Analysis Requests",
            "type": "integer"
          },
          "max_role_requests": {
            "default": 24,
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Max Role Requests",
            "type": "integer"
          }
        },
        "title": "Budget",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PauseReason": {
        "description": "暂停原因必须区分（PRD 5.2）。",
        "enum": [
          "NO_NEW_INFORMATION",
          "MANUAL",
          "PROVIDER_ERROR",
          "CONTEXT_LIMIT",
          "PROCESS_INTERRUPT"
        ],
        "title": "PauseReason",
        "type": "string"
      },
...
      "RunState": {
        "description": "会话运行状态（PRD 5.2）。没有其他状态。",
        "enum": [
          "READY",
          "RUNNING",
          "PAUSING",
          "STOPPING",
          "PAUSED",
          "ENDED"
        ],
        "title": "RunState",
        "type": "string"
      },
...
      "Scene": {
        "additionalProperties": false,
        "description": "场景／会话（PRD 3.1、5.2）。一个活动场景，多份历史记录。",
        "properties": {
          "background": {
            "title": "Background",
            "type": "string"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "pause_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PauseReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunState",
            "default": "READY"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "scene_id",
          "title",
          "background",
          "budget",
          "created_at"
        ],
        "title": "Scene",
        "type": "object"
      },
      "SceneAgent": {
        "additionalProperties": false,
        "description": "本场角色实例：``agent_id`` 是关联主键，``name`` 是显示字段。",
        "properties": {
          "agent_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "order_index": {
            "minimum": 0.0,
            "title": "Order Index",
            "type": "integer"
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/AgentSnapshot"
          }
        },
        "required": [
          "agent_id",
          "scene_id",
          "name",
          "order_index",
          "snapshot",
          "created_at"
        ],
        "title": "SceneAgent",
        "type": "object"
      },
      "SceneCreateRequest": {
        "additionalProperties": false,
        "description": "创建场景（会话）。\n\n``agents`` 的 2～8 人约束在契约层即校验，因此越界请求得到 422 而不是\n进入业务层（PRD 1.2、9）。",
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentSpecRequest"
            },
            "maxItems": 8,
            "minItems": 2,
            "title": "Agents",
            "type": "array"
          },
          "background": {
            "default": "",
            "maxLength": 2000,
            "title": "Background",
            "type": "string"
          },
          "max_analysis_requests": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Analysis Requests"
          },
          "max_role_requests": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Role Requests"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "agents"
        ],
        "title": "SceneCreateRequest",
        "type": "object"
      },
      "SceneDetailView": {
        "additionalProperties": false,
        "description": "场景详情：场景 + 本场角色集合（含模板快照）。",
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/SceneAgent"
            },
            "title": "Agents",
            "type": "array"
          },
          "locked": {
            "title": "Locked",
            "type": "boolean"
          },
          "scene": {
            "$ref": "#/components/schemas/Scene"
          }
        },
        "required": [
          "scene",
          "agents",
          "locked"
        ],
        "title": "SceneDetailView",
        "type": "object"
      },
      "SceneListView": {
        "additionalProperties": false,
        "properties": {
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/SceneSummaryView"
            },
            "title": "Scenes",
            "type": "array"
          }
        },
        "required": [
          "scenes"
        ],
        "title": "SceneListView",
        "type": "object"
      },
      "SceneSummaryView": {
        "additionalProperties": false,
        "description": "场景列表项。",
        "properties": {
          "agent_count": {
            "minimum": 0.0,
            "title": "Agent Count",
            "type": "integer"
          },
          "budget_locked": {
            "title": "Budget Locked",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "scene_id",
          "title",
          "status",
          "agent_count",
          "budget_locked",
          "created_at"
        ],
        "title": "SceneSummaryView",
        "type": "object"
      },
...
      "Usage": {
        "additionalProperties": false,
        "description": "用量。以服务端返回为准，**缺失记为 unknown，不是 0**（PRD 5.3）。\n\n本期不承诺货币费用；只记录原始用量，缓存命中字段在返回时记录。",
        "properties": {
          "cached_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached Tokens"
          },
          "input_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          }
        },
        "title": "Usage",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SceneWeave（角色互动沙盒）后端。M00 提供工程基础与公共契约；M01 提供角色模板与场景配置；M02/M03 提供可见性、调度与模型适配；M04 提供运行控制、事件、时间线、角色视角与 SSE；M06 提供只读行为分析。",
    "title": "SceneWeave",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scenes": {
      "get": {
        "operationId": "list_scenes_api_scenes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneListView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenes",
        "tags": [
          "m01-scenes"
        ]
      },
      "post": {
        "operationId": "create_scene_api_scenes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scene",
        "tags": [
          "m01-scenes"
        ]
      }
    },
...
    "/api/scenes/{scene_id}/analyses": {
      "get": {
        "description": "列出分析记录，并把操作数与实际模型请求数分开呈现（PRD 5.3）。",
        "operationId": "list_analyses_api_scenes__scene_id__analyses_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Analyses",
        "tags": [
          "m06-analysis"
        ]
      },
      "post": {
        "description": "执行一次分析。\n\n结果有三种“没有真正调用模型”的可能：`BLOCKED`（本地边界规则，含超长与预算耗尽）、\n`DISABLED`（能力关闭或外部包不可用）；两者 `provider_attempts=0`。",
        "operationId": "create_analysis_api_scenes__scene_id__analyses_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRecordView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Analysis",
        "tags": [
          "m06-analysis"
        ]
      }
    },
    "/api/scenes/{scene_id}/analyses/capability": {
      "get": {
        "description": "分析能力是否可用，以及不可用的具体原因。",
        "operationId": "analysis_capability_api_scenes__scene_id__analyses_capability_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisCapabilityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analysis Capability",
        "tags": [
          "m06-analysis"
        ]
      }
    },
...
  }
}